First configuration for the grid
The configuration allows the image to be placed on the grid and it takes a maximum size of 900 pixels or adapts if it is smaller.
The configuration allows you to select cells from the grid in the color black.
Still other colors, sizes, or being able to reverse cells are not allowed